Add FeedGrid and Sidebar components
