+++ src/components/FeedGridItem/FeedGridItem.js
@@ -1,83 +1,16 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import withHandlers from 'recompose/withHandlers';
-import { connect } from 'react-redux';
-import compose from 'recompose/compose';
-import Icon from '../Icon/Icon';
+import Icon from '../SimpleIcon/SimpleIcon';
 import formatDate from '../../utils/formatDate';
 import { getFirstParagraph } from '../../utils/getFirstParagraph';
-import * as actions from '../../actions';
-import { isFeedItemRead, isFeedItemBookmarked } from '../../ducks';
-
-const enhance = compose(
-  connect(
-    (state, ownProps) => ({
-      readOnOpen: state.settings.readOnOpen,
-      showSummary: state.settings.showSummary,
-      showContent: state.settings.showContent,
-      showImages: state.settings.showImages,
-      isBookmarked: isFeedItemBookmarked(
-        state,
-        ownProps.subscriptionId,
-        ownProps.item.id,
-      ),
-      isMarkedRead: isFeedItemRead(
-        state,
-        ownProps.subscriptionId,
-        ownProps.item.id,
-      ),
-    }),
-    {
-      markAsRead: actions.markAsRead,
-      unmarkAsRead: actions.unmarkAsRead,
-      updateUnreadCount: actions.updateUnreadCount,
-      unbookmarkFeedItem: actions.unbookmarkFeedItem,
-      bookmarkFeedItem: actions.bookmarkFeedItem,
-    },
-  ),
-  withHandlers({
-    onToggleMarkedAsRead: props => () => {
-      if (props.isMarkedRead) {
-        props.unmarkAsRead(props.subscriptionId, props.item.id);
-      } else {
-        props.markAsRead(props.subscriptionId, props.item.id);
-      }
-      props.updateUnreadCount(props.subscriptionId);
-    },
-    onToggleBookmarked: props => () => {
-      if (props.isBookmarked) {
-        props.unbookmarkFeedItem(props.subscriptionId, props.item.id);
-      } else {
-        props.bookmarkFeedItem(props.subscriptionId, props.item.id);
-      }
-    },
-    onOpenItemLink: props => e => {
-      if (props.readOnOpen) {
-        props.markAsRead(props.subscriptionId, props.item.id);
-        props.updateUnreadCount(props.subscriptionId);
-      }
-    },
-    onOpenLinksInNewTab: props => e => {
-      e.preventDefault();
-
-      if (e.target.href) {
-        window.open(e.target.href, '_blank');
-      }
-
-      if (e.target.parentNode.href) {
-        window.open(e.target.parentNode.href, '_blank');
-      }
-    },
-  }),
-);
 
 export const ListItem = props => (
   <li
-    className={`list-item
-    ${props.isMarkedRead ? 'list-item--marked-read' : ''}
-    ${props.showImages ? '' : 'list-item--hide-images'}`}>
-    <span className="list-item__date">{formatDate(props.item.pubDate)}</span>
-    <h3 className="list-item__title">
+    className={`feed-grid-item
+    ${props.isMarkedRead ? 'feed-grid-item--marked-read' : ''}
+    ${props.showImages ? '' : 'feed-grid-item--hide-images'}`}>
+    <span className="feed-grid-item__date">{formatDate(props.item.pubDate)}</span>
+    <h3 className="feed-grid-item__title">
       <a
         href={props.item.url}
         target="_blank"
@@ -89,7 +22,7 @@ export const ListItem = props => (
     {!props.isMarkedRead &&
       props.showSummary && (
         <p
-          className="list-item__description"
+          className="feed-grid-item__description"
           onClick={props.onOpenLinksInNewTab}
           dangerouslySetInnerHTML={{ __html: props.item.description }}
         />
@@ -97,7 +30,7 @@ export const ListItem = props => (
     {!props.isMarkedRead &&
       props.showContent !== 'none' && (
         <p
-          className="list-item__content"
+          className="feed-grid-item__content"
           onClick={props.onOpenLinksInNewTab}
           dangerouslySetInnerHTML={{
             __html:
@@ -111,7 +44,7 @@ export const ListItem = props => (
       href={props.item.url}
       target="_blank"
       rel="noreferrer noopener"
-      className="list-item__link"
+      className="feed-grid-item__link"
       onClick={props.onOpenItemLink}>
       <Icon name="link" size="small" title={props.item.title} />
     </a>
@@ -155,4 +88,4 @@ ListItem.propTypes = {
   isBookmarked: PropTypes.bool,
 };
 
-export default enhance(ListItem);
+export default (ListItem);
