+++ src/components/Loading/index.js
@@ -1,0 +1,1 @@
+export { default } from './Loading';
