+++ src/components/SettingsConfig/styles.scss
