+++ src/components/SimpleIcon/styles.scss
