+++ src/routes/App.js
@@ -4,12 +4,13 @@ import { connect } from 'react-redux';
 import compose from 'recompose/compose';
 import lifecycle from 'recompose/lifecycle';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
-import Subscriptions from '../Subscriptions';
-import Settings from '../Settings/Settings';
-import SidebarList from '../../components/SidebarList/SidebarList';
-import { COLOUR_MAP } from '../../data';
-import './styles.css';
-import * as actions from '../../actions';
+import Subscriptions from './Subscriptions/index';
+import Settings from './Settings/Settings';
+import Sidebar from '../components/Sidebar';
+import { COLOUR_MAP } from '../data/index';
+import * as actions from '../actions/index';
+
+import './App.css';
 
 const enhance = compose(
   connect(state => ({ settings: state.settings }), { saveSettings: actions.saveSettings }),
@@ -25,7 +26,7 @@ const App = ({ accentColour }) => (
   <Router>
     <div className="app">
       <div className="app-sidebar">
-        <SidebarList />
+        <Sidebar />
       </div>
       <div className="app-content">
         <div className="container">
