+++ src/routes/Subscriptions/Bookmarked.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import SubscriptionPreview from '../../components/SubscriptionPreview/SubscriptionPreview';
+import FeedGrid from '../../components/FeedGrid';
 import { getBookmarkedItemObj, getSubscriptionKeys } from '../../ducks';
 
 const enhance = connect(
@@ -14,7 +14,7 @@ const enhance = connect(
 const Bookmarked = props => (
   <div>
     {props.subscriptionIds.map(id => props.bookmarkedItems[id].length > 0 ? (
-      <SubscriptionPreview id={id} feedItems={props.bookmarkedItems[id]} key={id} />
+      <FeedGrid id={id} feedItems={props.bookmarkedItems[id]} key={id} />
     ) : (
       null
     ))}
